Completed mediapipe + opencv implementation
it doesn't run on the virtual device on my android studio, but works perfectly fine when tested on a physical OPPO device.

Has a bug where when the screen is rotated the app will SOMETIMES crash. Only sometimes though, yet to do testing to figure out what is causing it.